+++ book-exchange-backend/src/app.ts
@@ -3,7 +3,9 @@ import { testConnection } from "./services/db"; // Adjust the path as needed
 // import authRoutes from "./routes/legacy/authRoute";
 import authRoutes from "./routes/userAuthRoutes";
 import userManageRoutes from "./routes/userManagementRoutes"
-import bookRoutes from "./routes/legacy/booksRoute";
+// import bookRoutes from "./routes/legacy/booksRoute";
+import bookRoutes from "./routes/bookManagementRoutes";
+
 import userActionRoutes from "./routes/legacy/userActionRoute";
 import userRoutes from "./routes/legacy/userRoute"
 import { corsMiddleware } from "./middlewares/corsMiddleware";
@@ -25,7 +27,6 @@ app.get("/test-db", async (req, res) => {
 // Routes
 app.use("/api/auth", authRoutes);
 app.use("/api/users", userManageRoutes);
-
 app.use("/api/books", bookRoutes);
 app.use("/api/action", userActionRoutes);
 app.use("/api", userRoutes);